+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Please only add new entries below the [Unreleased](#unreleased---releasedate) he
 ### Features
 
 - **core**: Added `Location` provider to manage and track user navigation positions within application windows. (#740 @M-Adoo)
+- **widgets**: Added `Router` widget to handle navigation within window. (#746 @M-Adoo)
 
 ### Fixed
 - **core**: Fixed `Reuseable` panic when when holding the same Reuseable while pipe changes continuously.(#741 @wjian23)